exercicios aula 26

+++ quinzena7/labenusers/src/App.js
@@ -4,57 +4,48 @@ import axios from 'axios';
 import RegisterPage from './components/RegisterPage';
 import UsersPage from './components/UsersPage';
 
+
 const MainContainer = styled.div`
+  color: #FFF;
   text-align: center;
   background-color: #282c34;
   min-height: 100vh;
-`
 
-const Title = styled.h1`
-  color: white;
-  font-size: calc(10px + 2vmin);
-  margin-top: 0;
-`
-
-const DivInput = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  input{
-    margin: 0 5px;
+  button{
+    cursor: pointer;
   }
 `
 
 export default class App extends React.Component {
 
   state = {
-    currentPage: 'home',
+    currentPage: 'RegisterPage',
     usersList: [],
+    userData: '',
     inputName: '',
     inputEmail: ''
   }
 
   onChangeInputName = (event) => {
-    this.setState({ 
+    this.setState({
       inputName: event.target.value
     })
   }
 
   onChangeInputEmail = (event) => {
-    this.setState({ 
+    this.setState({
       inputEmail: event.target.value
     })
   }
 
   createUser = () => {
 
-    if (this.state.inputName.length > 2 && this.state.inputEmail.length > 5 && this.state.inputEmail.includes('@')){
+    if (this.state.inputName.length > 2 && this.state.inputEmail.length > 5 && this.state.inputEmail.includes('@')) {
       const body = {
         name: this.state.inputName,
         email: this.state.inputEmail
       }
-  
+
       axios.post("https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users", body, {
         headers: {
           Authorization: "rafael-rosa-munoz"
@@ -65,8 +56,8 @@ export default class App extends React.Component {
         alert('Erro ao cadastrar o usuário')
       })
 
-      this.setState({ inputName: '', inputEmail: ''})
-    }else {
+      this.setState({ inputName: '', inputEmail: '' })
+    } else {
       alert('Por favor, verifique os campos e preencha corretamente.')
     }
   }
@@ -78,23 +69,41 @@ export default class App extends React.Component {
         Authorization: "rafael-rosa-munoz"
       }
     }).then((response) => {
-      this.setState({usersList: response.data})
+      this.setState({ usersList: response.data })
     }).catch((error) => {
       console.log(error.message)
     })
   }
 
-  deleteUser = (id) => {
+  getUserData = (idUser) => {
 
-    axios.delete(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`, {
+    axios.get(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${idUser}`, {
       headers: {
         Authorization: "rafael-rosa-munoz"
       }
     }).then((response) => {
-      this.getAllUsers()
+      this.setState({ userData: response.data })
+      this.showUserDetails(response.data.name, response.data.email)
     }).catch((error) => {
       console.log(error.message)
     })
+  }
+
+  deleteUser = (id, name) => {
+
+    const confirmation = window.confirm("Você tem certeza que deseja deletar " + name + " da lista?")
+
+    if (confirmation) {
+      axios.delete(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`, {
+        headers: {
+          Authorization: "rafael-rosa-munoz"
+        }
+      }).then((response) => {
+        this.getAllUsers()
+      }).catch((error) => {
+        console.log(error.message)
+      })
+    }
   }
 
   onClickChangePage = (page) => {
@@ -102,32 +111,38 @@ export default class App extends React.Component {
       currentPage: page
     })
   }
-  
-  
-  render(){
+
+  showUserDetails = (name, email) => {
+    alert("Nome: " + name + "\nEmail: " + email)
+  }
+
+
+  render() {
 
     return (
       <MainContainer>
-        {/* <button onClick={this.getAllUsers}> Obter Usuários</button> */}
-        {this.state.currentPage === 'users' ? 
-        <UsersPage
-        onClickChangePage={this.onClickChangePage}
-        getAllUsers={this.getAllUsers}
-        usersList={this.state.usersList}
-        deleteUser={this.deleteUser}
-        ></UsersPage> : 
-        <RegisterPage
-        onClickChangePage={this.onClickChangePage}
-        valueInputName={this.state.inputName}
-        valueInputEmail={this.state.inputEmail}
-        onChangeInputName={this.onChangeInputName}
-        onChangeInputEmail={this.onChangeInputEmail}
-        createUser={this.createUser}
-        ></RegisterPage>
+        {
+          this.state.currentPage === 'UsersPage' ?
+            <UsersPage
+              onClickChangePage={this.onClickChangePage}
+              getAllUsers={this.getAllUsers}
+              usersList={this.state.usersList}
+              deleteUser={this.deleteUser}
+              getUserData={this.getUserData}
+            ></UsersPage> :
+            <RegisterPage
+              onClickChangePage={this.onClickChangePage}
+              valueInputName={this.state.inputName}
+              valueInputEmail={this.state.inputEmail}
+              onChangeInputName={this.onChangeInputName}
+              onChangeInputEmail={this.onChangeInputEmail}
+              createUser={this.createUser}
+            ></RegisterPage>
         }
       </MainContainer>
     );
   }
-  
+
 }
+
